HUD manager.
This adds a new class that manages displaying text to the player. It provides a user-friendly interface for showing categories of text, such as banner messages and input hints, and converts those into headers and subtitles accordingly.

Additionally, this includes some light cleanup of things that rely on system timing in order to tick at specific durations.