Early-out on zero/empty blocks in text rendering.
This adds checks to the text rendering ASM function to verify whether or not we have any blocks and whether or not those blocks have zero lines and, if not, aborts rather than crashing the game.

To support this, the AsmBuilder now has a smart bookmarking system for handling intra-function jump statements.